Changes in ChoreController, ChoreUI. Added Chore and ChoreList

+++ source/core/src/main/com/deco2800/game/screens/maingame/MainGameTextDisplay.java
@@ -76,9 +76,9 @@ public class MainGameTextDisplay extends UIComponent {
 
         // Display Text
         displayText = new Label("", skin, "large");
-        displayText.setSize(colWidth*6, rowHeight*3);
-        displayText.setPosition(colWidth*2, (float) rowHeight/2);
-        displayText.setFontScale((float) (colWidth*10)/1280); // Scale font to screen size
+        displayText.setSize(colWidth*6f, rowHeight*3f);
+        displayText.setPosition(colWidth*2f, rowHeight/2f);
+        displayText.setFontScale((colWidth*10f)/1280f); // Scale font to screen size
         displayText.setWrap(true);
 
         stage.addActor(displayText);
